Emergency quickfix in geoip2dat.py - a new field was inserted by Maxmind on CSV files.
Maxmind inserted a new field into the CSV files called "is_anycast", and this broke geoip2dat.py CSV reader. This change occurred around 5 days ago. Emergency insertion of the new field in the list of "fields" of the CSV reader that generates the .dat.gz files so that they can be updated.

I will shield the code so that this type of change will no longer affect us in the future. The new releases of GeoIP2Fast v1.2.2 and GeoIP2Fast v1.1.11 (Legacy) will be published in the next few days.

This did not affect the search. It has only affected the creation of new dat.gz files since 1/19/2024.

https://github.com/rabuchaim/geoip2fast/actions/runs/7589712513/job/20674920387

I will create an SMS alert in the github actions workflow to notify me immediately when an error occurs in the automatic creation of files.